Move PDF permission questions into own section
This improves upon PR #1044 by moving the permission questions into its own sub-section in the PDF right below the "Children’s relationship with parties" one, so it is a bit more clear.

Also because it is a separate section we can apply a bit of styling so the question column is much wider than the answer column (because answer can only be "Yes").

Apart from that, and to make things even more easy and clear to read, I've applied a bit of shading to the top-level section headers to separate further each of the parts of the form.
