+++ spec/presenters/summary/sections/children_relationships_spec.rb
@@ -7,16 +7,12 @@ module Summary
         applicants: applicants,
         respondents: respondents,
         other_parties: other_parties,
-        relationships: relationships_scope,
       )
     }
 
     let(:applicants) { double('applicants') }
     let(:respondents) { double('respondents') }
     let(:other_parties) { double('other_parties') }
-
-    let(:relationships_scope) { double('relationships', with_permission_data: relationships) }
-    let(:relationships) { [] }
 
     subject { described_class.new(c100_application) }
 
@@ -62,70 +58,6 @@ module Summary
 
         expect(answers[3]).to be_an_instance_of(Partial)
         expect(answers[3].name).to eq(:row_blank_space)
-      end
-
-      context 'when there are relationships with permission details' do
-        let(:relationships) { [relationship1, relationship2] }
-
-        let(:person) { instance_double(Applicant, full_name: 'Applicant Test') }
-        let(:minor)  { instance_double(Child, full_name: 'Child Test') }
-
-        let(:relationship1) {
-          instance_double(
-            Relationship,
-            person: person,
-            minor: minor,
-            parental_responsibility: 'no',
-            living_order: 'yes',
-            amendment: nil,
-            time_order: nil,
-            living_arrangement: nil,
-            consent: nil,
-            family: nil,
-            local_authority: nil,
-            relative: nil,
-          )
-        }
-
-        let(:relationship2) {
-          instance_double(
-            Relationship,
-            person: person,
-            minor: minor,
-            parental_responsibility: 'no',
-            living_order: 'no',
-            amendment: 'no',
-            time_order: 'yes',
-            living_arrangement: nil,
-            consent: nil,
-            family: nil,
-            local_authority: nil,
-            relative: nil,
-          )
-        }
-
-        it 'has the correct number of rows' do
-          expect(answers.count).to eq(7)
-        end
-
-        it 'has the correct rows in the right order' do
-          expect(answers[0]).to be_an_instance_of(FreeTextAnswer)
-          expect(answers[0].question).to eq(:applicants_relationships)
-          expect(answers[0].value).to eq('applicants_relationships')
-
-          expect(answers[1]).to be_an_instance_of(Answer)
-          expect(answers[1].question).to eq('child_permission_living_order')
-          expect(answers[1].value).to eq('yes')
-          expect(answers[1].i18n_opts).to eq({ applicant_name: 'Applicant Test', child_name: 'Child Test'})
-
-          expect(answers[2]).to be_an_instance_of(Answer)
-          expect(answers[2].question).to eq('child_permission_time_order')
-          expect(answers[2].value).to eq('yes')
-          expect(answers[2].i18n_opts).to eq({ applicant_name: 'Applicant Test', child_name: 'Child Test'})
-
-          expect(answers[3]).to be_an_instance_of(Partial)
-          expect(answers[3].name).to eq(:row_blank_space)
-        end
       end
     end
   end
